+++ .vscode/settings.json
@@ -1,3 +1,4 @@
 {
-  "liveServer.settings.port": 5501
+  "liveServer.settings.port": 5501,
+  "cSpell.words": ["anwsomebook", "luxon", "showlist"]
 }
