+++ index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="index.css" />
-    <script src="index.js" defer type="module"></script>
+    <script src="index.mjs" defer type="module"></script>
     <title>Awesome Books Title Page</title>
   </head>
