+++ modules/date.js
@@ -1,3 +1,11 @@
+import { DateTime } from 'luxon';
+
 export default function updateDate() {
-  document.querySelector('.date').innerHTML = new Date().toString();
+  const dateTime = DateTime.now();
+  const customFormat = dateTime.toFormat('MMMM dd, yyyy');
+  // display the formatted date in an HTML element
+  const element = document.querySelector('.date');
+  if (element !== null) {
+    element.textContent = customFormat;
+  }
 }
