+++ package.json
@@ -22,5 +22,6 @@
   },
   "dependencies": {
     "luxon": "^3.3.0"
-  }
+  },
+  "type": "module"
 }
